fix(web): co-locate details placeholder, parameterize usage meter count
Review findings on Task 2's skeleton placeholders:
- move renderDetailsPlaceholderHtml next to its real counterpart
  renderDetailsHtml in workspace-rail.ts, so they stay editable in view
  of each other as the module doc requires
- give renderUsagePlaceholderHtml a meters param (default 2) instead of
  hard-coding two rows, since renderUsageHtml can render two meters, one
  meter, or a capless text line depending on the workspace's plan
